refactor(agents): repoint consolidated tool imports
